Move desktop to internal
