+++ internal/desktop/desktop_platform_darwin.go
