+++ internal/desktop/icon_windows.syso
