+++ internal/desktop/systray_linux.go
