+++ internal/desktop/systray_nonlinux.go
